Update README.md
This project includes three Tableau dashboards analyzing the impact of COVID-19 worldwide and in Spain:

Global Overview: A high-level visualization of how COVID-19 affected different regions over time. Since this is an image, it is not interactive, but it was designed to show the progression of the pandemic.

Policy Impact Analysis: A comparative dashboard that groups countries based on their COVID-19 regulations (strict, moderate, and flexible) to analyze how these policies influenced the pandemic's development.

Spain Deep Dive: A focused analysis of COVID-19 in Spain, a country where the pandemic sparked significant controversy.

This project aims to provide insights into the pandemic’s evolution, the effectiveness of different policies, and a deeper look into Spain’s unique situation.